move files to make nextjs happy
