+++ playground/src/components/request_submission_selection.tsx
@@ -1,10 +1,10 @@
 import {useState} from "react";
-import { Exercise } from "@/pages/model/exercise";
-import ExerciseSelect from "@/pages/components/exercise_select";
-import ModuleResponse from "@/pages/model/module_response";
-import ModuleResponseView from "@/pages/components/module_response_view";
-import { Submission } from "@/pages/model/submission";
-import {ModuleMeta} from "@/pages/model/health_response";
+import { Exercise } from "@/model/exercise";
+import ExerciseSelect from "@/components/exercise_select";
+import ModuleResponse from "@/model/module_response";
+import ModuleResponseView from "@/components/module_response_view";
+import { Submission } from "@/model/submission";
+import {ModuleMeta} from "@/model/health_response";
 
 async function requestSubmissionSelection(athenaUrl: string, module: ModuleMeta, exercise: Exercise | null): Promise<ModuleResponse | undefined> {
     if (!exercise) {
