+++ playground/src/model/exercise.ts
